requirements.txt
solidpython